refactor: narrow epoch (from uint256 to uint64) and commision types (uint256 to unit16) in public interfaces
Affected Signatures:
- setPayeeMaxCommission(address token, uint16 maxBps)
- createRail(address token, address from, address to, address arbiter, uint16 commissionRateBps)
- modifyRailLockup(uint256 railId, uint64 period, uint256 lockupFixed)
- settleRail(uint256 railId, uint64 untilEpoch)
- getRail(uint256 railId) -> returns RailView {
    // narrowed:
    uint64 lockupPeriod;
    uint64 settledUpTo;
    uint64 endEpoch;
    uint16 commissionRateBps;
    // other fields unchanged
  }
- getRailsForPayerAndToken(address payer, address token) → returns RailInfo[] { uint64 endEpoch; other fields unchanged }
- getRailsForPayeeAndToken(address payee, address token) → returns RailInfo[] { uint64 endEpoch; other fields unchanged }
- RateChangeQueue.enqueue(Queue storage, uint256 rate, uint64 untilEpoch)

+++ src/Payments.sol
@@ -351,7 +351,7 @@ contract Payments is
     /// @param maxBps The maximum commission in basis points (0-10000).
     function setPayeeMaxCommission(
         address token,
-        uint256 maxBps
+        uint16 maxBps
     ) external validateNonZeroAddress(token, "token") {
         require(maxBps <= COMMISSION_MAX_BPS, "max commission exceeds maximum");
         payeeCommissionLimits[token][msg.sender] = PayeeCommissionLimit({
@@ -492,7 +492,7 @@ contract Payments is
         address from,
         address to,
         address arbiter,
-        uint256 commissionRateBps
+        uint16 commissionRateBps
     )
         external
         nonReentrant
@@ -532,7 +532,7 @@ contract Payments is
         rail.to = to;
         rail.operator = operator;
         rail.arbiter = arbiter;
-        rail.settledUpTo = block.number;
+        rail.settledUpTo = uint64(block.number);
         rail.endEpoch = 0;
         rail.commissionRateBps = commissionRateBps;
 
@@ -553,7 +553,7 @@ contract Payments is
     /// @custom:constraint Operator must have sufficient lockup allowance to cover any increases the lockup period or the fixed lockup.
     function modifyRailLockup(
         uint256 railId,
-        uint256 period,
+        uint64 period,
         uint256 lockupFixed
     )
         external
@@ -574,7 +574,7 @@ contract Payments is
 
     function modifyTerminatedRailLockup(
         Rail storage rail,
-        uint256 period,
+        uint64 period,
         uint256 lockupFixed
     ) internal {
         require(
@@ -609,7 +609,7 @@ contract Payments is
 
     function modifyNonTerminatedRailLockup(
         Rail storage rail,
-        uint256 period,
+        uint64 period,
         uint256 lockupFixed
     ) internal {
         Account storage payer = accounts[rail.token][rail.from];
@@ -769,7 +769,7 @@ contract Payments is
         uint256 newRate,
         uint256 oneTimePayment
     ) internal {
-        uint256 endEpoch = maxSettlementEpochForTerminatedRail(rail);
+        uint64 endEpoch = maxSettlementEpochForTerminatedRail(rail);
         require(
             newRate == 0 && oneTimePayment == 0,
             "for terminated rails beyond last settlement epoch, both new rate and one-time payment must be 0"
@@ -824,7 +824,7 @@ contract Payments is
             // For arbitrated rails, we need to enqueue the old rate.
             // This ensures that the old rate is applied up to and including the current block.
             // The new rate will be applicable starting from the next block.
-            rail.rateChangeQueue.enqueue(oldRate, block.number);
+            rail.rateChangeQueue.enqueue(oldRate, uint64(block.number));
         }
     }
 
@@ -929,12 +929,12 @@ contract Payments is
             uint256 totalNetPayeeAmount,
             uint256 totalPaymentFee,
             uint256 totalOperatorCommission,
-            uint256 finalSettledEpoch,
+            uint64 finalSettledEpoch,
             string memory note
         )
     {
         // Verify the current epoch is greater than the max settlement epoch
-        uint256 maxSettleEpoch = maxSettlementEpochForTerminatedRail(
+        uint64 maxSettleEpoch = maxSettlementEpochForTerminatedRail(
             rails[railId]
         );
         require(
@@ -956,7 +956,7 @@ contract Payments is
     /// @return note Additional information about the settlement (especially from arbitration).
     function settleRail(
         uint256 railId,
-        uint256 untilEpoch
+        uint64 untilEpoch
     )
         public
         nonReentrant
@@ -968,7 +968,7 @@ contract Payments is
             uint256 totalNetPayeeAmount,
             uint256 totalPaymentFee,
             uint256 totalOperatorCommission,
-            uint256 finalSettledEpoch,
+            uint64 finalSettledEpoch,
             string memory note
         )
     {
